pyproject.toml
[tool.poetry]
name = "wanted"
version = "0.1.0"
description = ""
authors = ["SonJinHYo <wlsgy95@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "4.2.0"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
